idk... started to make mesh class based on the LearnOpenGl tutorial but quickly realized shit was getting real.
so, now I should maybe start to get the game stuff going. a lot of things, like models and light sources need (for me, I guess?) to make sense in context of something else.

in other words, if I want to have a rendering engine with an API, I need to have some kind of use case for that API. So, I'll start skeletoning it out.