AGHHH... the material.shininess was never set... so there was some NaN going on or something....
fixed that by "jerry-rigging" some bs system. turns out that Assimp splits up Models into their meshes where each mesh has only one material! So, I made a member for the material's shininess on the mesh, which I calculated before construction in Model by averaging the AI_MATKEY_SHININESS values of each aiMaterial that was am aiTextureType_SPECULAR. Awesome. The code is kinda ugly and embarassing, but it works for now.

The camera is still messed up, though. Thinking of initializing that with the window.